+++ changelog.md
@@ -19,12 +19,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gdi.fill_solid_rect
 - kernel.QueryInformationJobObject
 - sample timeit.py
+- sample dump_proc_env.py
 
 ### Changed
 
 - let sample 'restart_usb_port.py' use cfgmgr.CM_Get_Device_Interface_List
   instead of setupapi.SetupDiGetDeviceInterfaceDetail
 - support for addition and subtraction of FILETIME
+- add parameter 'inherit' to CreateEnvironmentBlock and create_env_block_as_dict
+
+### Fixed
+
+- fixed handling of "\0\0" in ctwin32.multi_str_from_str
 
 ## [2.5.0] - 2023-03-11